Add a regression test suite
It runs autodescribe and fv against a bunch of test files and compares
the output with what is expected. Not all test files in the corpus are
currently used, but they are all supported by at least one program in
fileviewinfo (or will be shortly) and can be useful for manual and
future automated tests.