Make more file types safe for files with dash-prefixed names
Many programs will treat a file name starting with a dash as
introducing command-line options. Disable this where possible to avoid
undesired results when such a file is given. This is not done when core
system programs are called (like ar) for portability, since they don't
all support this. Also, switch to pamfile from pnmfile in one place
since the latter is deprecated.