Improve STL handling in autodescribe & fvi
Strip trailing whitespace, including \r characters which are sometimes
found in these files. Decode the comments as being in the CP1252
character set, which is not always correct but is better than letting
the terminal try to handle them as-is or handling them as ISO 8859/1.
Improve detection of ASCII vs binary STL format, which was previously
confused by descriptions starting with "solid", which is surprisingly
common.