Improve portability
Some sed commands are fixed to work on FreeBSD. A couple of comments now
mention when specific versions of some programs are required.
Specifically, the FreeBSD version of unzip does not provide a way to
extract the archive comment, and GNU tar requires a --wildcards argument
to extract using a wildcard pattern whereas FreeBSD does not (and we
don't do auto-detection of which version is available).