Eliminate problems with files starting with dashes
In some cases, the file names would be interpreted as command-line
options to programs and could, in the worst case, cause the programs to
delete, overwrite or corrupt files. If the helper program doesn't
support a way to treat a dashed file as a file and not an option then
prepend a ./ to the name to prevent it from causing problems.