Add support for CMakeFiles.txt files to autodescribe
Support is also added for embedded CMake files within tarballs.
The test file uses a new convention of exttypeN.T where T is the type
designator as used in the -t option. This is for test files that can
only be automatically detected when given a magic file name instead of
a simple known file extension.