Handle &lt; and &gt; for HTML in autodescribe
These are probably the two most common character entities, so might as
well convert them. Using an HTML parser would be better (or calibre) but
the simple text scanning method for HTML is much more likely to have all
the required dependencies available, and works just as well in most real
world cases.